requirements.txt
matplotlib
altair
vl-convert-python
pandas>1.2
numpy
docopt>0.6.1
PyYAML>5.4
selenium<4.3
linkify-it-py
problem_bank_scripts
